+++ app/src/main/java/me/juangoncalves/mentra/db/models/PortfolioValueModel.kt
@@ -3,10 +3,11 @@ package me.juangoncalves.mentra.db.models
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Entity(tableName = "PortfolioValue")
 data class PortfolioValueModel(
-    @ColumnInfo(name = "usd_value") val valueInUSD: Double,
+    @ColumnInfo(name = "usd_value") val valueInUSD: BigDecimal,
     @PrimaryKey val date: LocalDate = LocalDate.now()
 )
